Make error reporting in SQLite3 backend more flexible
Allow passing a lambda instead of just a hard-coded error message to
execude_hardcoded() helper function, so that we could include additional
information in the error messages without necessarily constructing the
error message in advance -- which is useless if it's not used at all,
i.e. if the error doesn't happen.

No real changes yet, but this prepares ground for the next commit.