Find MariaDB headers/libraries when looking for MySQL
So far this is only done under non-Windows systems, as this is where
MariaDB is mostly found.

Also simplify the paths by removing duplicates and don't bother looking
in some ad hoc directories by default, MYSQL_DIR can always be set
explicitly if MySQL is available in a non-standard location.