Revamp Firebird Blob backend implementation
Besides some general code cleanup here and there, it has been ensured
that blob objects are not trimmed to zero once they have been used. This
allows to reuse the same blob object to insert the same data into
multiple rows or tables (as individual Blob objects in the DB).