Move check for PostgreSQL 9.3 to CMake from the source file
Unfortunately pg_config.h is not available in all installations under
Windows, so don't rely on being able to include it and instead check
POSTGRESQL_VERSION in CMake and define SOCI_POSTGRESQL_NO_LO64 if
necessary.

This has an additional advantage of allowing to test this code easily
even when using PostgreSQL >= 9.3 by turning on this option.