Work around bug in VS2015
There seems to be a bug in Visual Studio 2015 that leads to a call tp
logger_impl::set_stream instead of standard_logger_impl::set_stream in
the "basic logging support" test case, which throws a "not supported"
exception instead of actually setting the log stream, leading to a test
case failure. This hasn't been observed for any other compiler,
including more recent versions of Visual Studio. Thus, it was concluded
that this was indeed a compiler bug in VS2015.

Moving the standard_logger_impl implementation out of the anonymous
namespace works around this bug without affecting the behavior of the
library in any way.