Install PostgreSQL ourselves in macOS CI workflow
It used to be part of macos-12 image but is not present in the later
ones for whatever reason.

Installing postgresql formula using brew already initializes the
database cluster but does it in a non-default directory, so adjust
PGDATA definition accordingly and remove "pg_ctl init" which is not
needed any longer.