Initialize most session members in their declaration
This ensures that they are initialized, which previously wasn't the
case, e.g. gotData_ wasn't initialized in the default ctor and
poolPosition_ wasn't initialized at all.

This didn't seem to result in any real problems, but triggered UBSAN
reports when copying "bool gotData_" values which didn't have a valid
value for bool type.