Avoid accessing uninitialized type variable in Oracle classes
oracle_standard_into_type_backend::type_ may not be initialized when
clean_up() is called, and comparing its value with x_xmltype and other
constants was triggering undefined behaviour in this case, so avoid
doing it by checking if ociData_ is non-null before doing it.

Also modify oracle_standard_use_type_backend in a similar way, even if
it doesn't suffer from this problem, for symmetry.