Remove statement::define_for_row() and call it from describe()
It looks like this function must always be called after a (successful)
call to describe() and it is not called from anywhere else, so it seems
wrong to both require and allow calling it independently -- just call it
from describe() itself.

No real changes, this function has never been part of public API, so
hopefully nobody was using it.