Make query context caching configurable
The default is to cache parameters for all queries. This can be changed
to either only cache in case of an error or to never cache. The latter
will also ensure that the parameters will not be part of exception
messages in order to not leak any potentially sensitive information.