Allow to row::get blob into suitable containers
"Suitable" means that the container uses contiguous storage in memory
and holds entries of type T where sizeof(T) == 1. If this is given, then
the blob's data can be copied into a new instance of such a container,
which is then returned in the row::get<>() call.

This restores the previous possibility of reading blobs into objects of
type std::string but also generalizes the concept to any suitable
container (such as std::vector<unsigned char>).

Note: The check for whether the given container object uses contiguous
object storage is based on its iterators being classified as
random-access iterators. This is not a perfect check as in principle
there could be random-access iterators that don't iterate a contiguous
memory region. However, for typical containers this is very unexpected
as this would typically lead to rather inefficient access patterns when
making use of the random-access capability.
A container that is wrongly determined to be contiguous will likely
result in a program crash due to a segfault.

Fixes #1173