Work around bug in int8_t support when using SQL Server
Negative values saved to the database are stored as positive numbers in
at least some versions of the database, so even if they're converted
back to the negative ones when we read them back, ordering the result
set by them doesn't work as expected, see #1193.

Until this can be fixed, work around this by sorting the values in the
test itself instead.