Fix get_affected_rows() for partial updates in SQLite3 backend
This function always returned 0 even if some rows were updated because
rowsAffectedBulkTemp was never copied into rowsAffectedBulk_ in this
case, i.e. when load_one() threw an exception.

Fix this by just getting rid of this temporary variable and updating
rowsAffectedBulk_ directly, this seems to be much simpler than this
code, added back in c97a666e (SQLite3 support for get_affected_rows() in
bulk operations., 2013-03-11), and is more correct.