Implement getting parameter names for Oracle too
This was the only backend which didn't implement retrieving parameter
names by position, so this commit makes it possible to rely on this
working for all backends and will simplify writing upcoming tests.

Note that Oracle always returns names in upper case and there doesn't
seem to be any way to avoid it.