Add oracle_statement_backend::get_statement_attr() helper
Refactor the code to replace all calls to OCIAttrGet(OCI_HTYPE_STMT)
with calls to the new function. This is simpler and shorter.

Note that this is still unsafe as the correspondence between the
attribute and the corresponding type is not checked, but it doesn't seem
worth to complicate things further as long as we're using only a handful
of attributes (and each of them is used only once).

This commit is best viewed ignoring whitespace-only changes.