Get rid of unneeded rowsAffectedBulkTemp in PostgreSQL backend
Just use rowsAffectedBulk_ normally, there doesn't seem to be any need
for this temporary variable and using it is fragile as it's easy to
forget to copy it into rowsAffectedBulk_ itself before returning.

This is similar to what was done in 42d5a2ec (Fix get_affected_rows()
for partial updates in SQLite3 backend, 2025-01-28) for SQLite.