Improve FindMySQL.cmake
The improved version should make sure to always print a found message
(unless QUIET) is specified and it should respect externally specified
MySQL_INCLUDE_DIRS etc. variables in cases in which MySQL is not found
via vcpkg or pkg-config.

The weird dance with SOCI_MYSQL_DIRECT_INCLUDE has also been dropped,
assuming that the include path is always configured to allow a direct
include of the mysql.h header.