Transform MySQL_INCLUDE_DIRS only when using config program
This is not necessary in the other cases, such as when using pkg-config.

It's also not necessary to do it for MySQL_LIBRARIES as the arguments to
target_link_libraries() can be arbitrary linker options too.

Finally, make the regex more precise and only remove "-I" and, also,
"-isystem", instead of anything starting with "-".