Build SQLite3 ourselves if it's not otherwise available
Use sources from the submodule to create soci_sqlite3_builtin library.
Recommend cloning with --recurse-submodule to ensure that the submodule
is available and update AppVeyor CI to initialize submodules too.

Add "BUILTIN" dependency mode to soci_define_backend_target() and use it
for SQLite3 backend.

Note that because SQLite3 source is C, and not C++, this language needs
to be explicitly added to CMake project() function, otherwise C files
are just silently ignored.

This commit is best viewed ignoring whitespace-only changes.