Stop using soci_public_dependency() for Boost dependencies
This doesn't work when the call to it can be skipped due to WITH_BOOST
being FALSE and instead of making this function even more complicated,
it is simpler to just avoid using it here entirely (and simplify its
code by dropping support for arguments only needed to support this use
case in the next commit).

This means that SOCI config file doesn't define Boost targets any more,
but, first, it's not really clear whether it should do it (SOCI is
perfectly usable without Boost) and, second, even it it should, it would
be much simpler and more understandable to add special support for Boost
to soci-config.cmake instead of trying to handle in a generic function.