Simplify and centralize SQLite3 backend error generation
Call sqlite3_errmsg() directly in sqlite3_soci_error ctor instead of
doing it in multiple places in the backend code.

Throw soci_error when there is no SQLite3-specific information
associated with the error.

Update the documentation to make it more clear that both
sqlite3_soci_error and soci_error must be caught.