Build SQLite3 as part of soci_sqlite3 itself if necessary
Get rid of a separate soci_sqlite3_builtin target as it didn't work
correctly when using static libraries.

Instead, just compile the amalgamated SQLite3 source as part of
soci_sqlite3 itself, which works after the changes of the parent commit
as we don't get tons of warnings for this C file any longer.