Make Windows GitHub Actions workflow work
Fix the workflow which was added in master to make it possible to use
it, but which didn't work yet.

Build SOCI under Windows using MSVC and run the tests that we can easily
run in the GitHub Actions environment (to run the other ones we would
need to set up PostgreSQL ODBC drivers and MySQL and SQL Server
databases ourselves).