Get rid of unnecessary SOCI_ALLOW_DEPRECATED_BEGIN/END macros
They were ugly and not really necessary when we can just avoid marking
the functions as deprecated when compiling SOCI itself to allow using
them from the implementation code.

Also simply remove deprecated attributes from the functions in
blob_backend which is not supposed to be used publicly at all.