Output SOCI version and build type information at the end
Version is probably not very useful, but it is important to see at a
glance if SOCI is to be built as static or shared library.

Also move the status message about enabled backends to the end as well
to make it easier to find and appear even when re-configuring in the
same directory without any changes.