Implement tcp_user_timeout support with libpq < 12 under Linux
In addition to providing fallback implementation of this option under
Windows, also do it under Linux when using libpq < 12, where this option
was added.

This allows to always specify this option in the connection string,
without worrying about it resulting in fatal error from PQconnectdb() on
older systems: in the worst case, the option is ignored and when using
Linux (or any other system with TCP_USER_TIMEOUT support, although none
seem to currently exist) it will work in the same way as in libpq.