Fix truncating wide strings when using NVARCHAR in ODBC backend
ODBC SQLDescribeCol() returns column size in characters, but
SQLBindCol() expects the buffer size in bytes. The existing code failed
to multiply by sizeof(SQLWCHAR), which could result in buffer
truncation, corrupted reads, or subtle bugs when binding to wide
character types such as NVARCHAR.

Fix this by using the size in bytes.

Also update an existing unit test to show the problem: by simply
reducing the buffer size, it could be made to fail without the changes
in this commit.

Closes #1290.

Closes #1292.