Warn against incorrect find_package(SOCI) usage
It can happen that users specify e.g. SOCI_MYSQL CMake variables to
select SOCI backends when using find_package, just as they would when
using add_subdirectory. However, this has no effect which will be very
confusing.

Therefore, we add a check for this in case no explicit components are
given in the find_package call and if we detect any of the SOCI_*
variables corresponding to backends are defined, we emit a warning
referring to the use of find_package(SOCI COMPONENTS …) instead.

Closes #1294.