Move external symbol declaration into global namepace
Even though symbols with C linkage don't really have a namespace,
declaring an 'extern "C"' symbol inside of a C++ namespace causes the
symbol to be inside that namespace, with the exception that the symbol
name must still be globally unique. This can cause issues where this
external symbol is linked to an actual C library.

See also https://stackoverflow.com/q/28996944

Fixes #1266