Fix possible access to uninitialized memory in ODBC backend
If the connection string is truncated, do not read beyond the end of the
buffer, as would happen in this case as the returned string length would
be bigger than the buffer size we passed to SQLDriverConnect().

As we can rely on the string being NUL-terminated, just assign it to
connection_string_ directly.