Fix exception with multiple batch operations in SQLite3 backend
Don't leave current_row_ with an invalid value after finishing a batch
operation, it could be reused, resulting in an exception due to the
index being invalid, when the statement is executed again.

Note that this was already done for the other backends using
current_row_ in their overridden get_row_to_dump() in a similar way
(Firebird and PostgreSQL), so they were not affected by this bug.

Closes #1308.