Fix ODBC handle leaks in case of unsuccessful connection
SQLHENV and SQLHDBC handles stored in odbc_session_backend member
variables were never freed if the ctor threw an exception because the
dtor, freeing them, was not called in this case, resulting in memory and
resource leaks.

Fix this by wrapping these handles in a simple struct freeing them
automatically, as members dtors are always called if they had been
constructed successfully.