Fix leak of SQLite3 connection handle in case of failure
If an error happened in sqlite3_session_backend ctor, the connection
variable was leaked, as sqlite3_session_backend dtor normally freeing it
was not called in this case.

Fix this by ensuring that we call sqlite3_close() if ctor exits
prematurely. Note that this implies removing call to sqlite3_close()
from check_sqlite_err(), as calling sqlite3_close() twice is invalid.